test/web-platform-tests/to-upstream/FileAPI/blob/Blob-size.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>blob.size</title>
<link rel="author" title="Timothy Gu" href="mailto:timothygu99@gmail.com">
<link rel="help" href="https://w3c.github.io/FileAPI/#dfn-size">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
"use strict";

test(() => {
  const blob = new Blob(["TEST"]);
  assert_equals(blob.size, 4, "Initial blob should have a size of 4");
});
</script>
